feat:console log
