+++ log/1
@@ -1,0 +1,1 @@
+{"path": "log_test.go:15(golog.TestDebug)", "serviceName": "test", "traceID": null, "msg": "info"}
