+++ log/1info.log
@@ -1,0 +1,2 @@
+%s
+string data
